test(awsdiscover): describe extractor pins by what they test, not by bundle number
Rename test file and test functions so they describe the per-type
extractor contracts they pin, rather than a build-number prefix:

- File: cloudcontrol_bundle14_test.go → cloudcontrol_extractors_test.go
- Tests:
  - TestBundle14_BackupSelectionExtractors → TestBackupSelectionConfig
  - TestBundle14_CognitoUserPoolExtractors → TestCognitoUserPoolConfig
  - TestBundle14_IAMInstanceProfileExtractors → TestIAMInstanceProfileConfig
  - TestBundle14_LambdaEventSourceMappingExtractors → TestLambdaEventSourceMappingConfig
  - TestBundle14_SSMParameterExtractors → TestSSMParameterConfig
  - TestBundle14_EmptyTagsExtractor → TestEmptyTagsExtractor_NonNilEmptyMap

Also strip "Bundle 14" prose references from arn_rules.go, arn_rules_test.go,
cloudcontrol_types.go, cloudcontrol_discoverer.go — replace with what the
text is actually about.

Fold in qa-professor MINOR findings while touching these files:
- NativeIDs assertions use reflect.DeepEqual against exact wantNative maps
  (catches a regression that stamps extra keys, e.g. accidentally adding
  `name` alongside `arn` on the cognito/iam configs).
- Tag assertions use reflect.DeepEqual against exact wantTags maps
  (a regression that pulls a phantom key through extractTagList no longer
  slips past `tags["env"] == "prod"`-style partial assertions).
- IAM instance profile and SSM parameter NameHint tests differentiate the
  prop value from the identifier so a wrong-key mutation
  (`InstanceProfileName` → `Name`, or `Name` → `Path`) actually changes
  output and fails. Also adds explicit empty-props fallback cases.
- Lambda/SSM Config tests now explicitly pin IsGlobal=false and
  SkipProjectTagFilter=false (Cognito/IAM/Backup pinned theirs already;
  this brings them to parity).
- SkipProjectTagFilter_BypassesRGTCache asserts `listCalls != 1` instead
  of `== 0`, catching a "called twice" regression where a refactor
  consults the cache, finds empty, then also falls back to List.

No production-code logic changes.
